Official quotes and refactoring
