+++ src/components/Quotes/styles.module.css
@@ -1,0 +1,3 @@
+.quotes h1 {
+  text-align: center;
+}
