+++ src/pages/index.js
@@ -4,6 +4,7 @@ import Link from '@docusaurus/Link';
 import useDocusaurusContext from '@docusaurus/useDocusaurusContext';
 import Layout from '@theme/Layout';
 import HomepageFeatures from '@site/src/components/HomepageFeatures';
+import Quotes from '@site/src/components/Quotes';
 
 import styles from './index.module.css';
 
@@ -44,64 +45,6 @@ function VideoEmbed() {
   );
 }
 
-const quotesData = [{
-  name: "James Snell",
-  title: "Node.js Technical Steering Committee",
-  quote: "Platformatic is a great example of how to build a great developer experience on top of a database.",
-  avatar: "https://avatars.githubusercontent.com/u/439929?s=460&v=4"
-}, {
-  name: "Feross Aboukhadijeh",
-  title: "Founder & CEO, Socket.dev",
-  quote: "Quote from Feross",
-  avatar: "https://avatars.githubusercontent.com/u/121766?s=460&v=4"
-}, {
-  name: "Charlie Robbins",
-  title: "",
-  quote: "Quote from Charlie",
-  avatar: "https://avatars.githubusercontent.com/u/4624?s=460&v=4"
-}, {
-  name: "David Mark Clements",
-  title: "",
-  quote: "Quote from Dave",
-  avatar: "https://avatars.githubusercontent.com/u/1190716?s=460&v=4"
-}]
-
-function Quotes() {
-  return (
-    <section className={styles.quotes}>
-      <div class="container">
-        <div class="row">
-          <div class="col col--12">
-            <h1>Quotes</h1>
-          </div>
-
-          {quotesData.map((quote, index) => ( 
-          <div class="col col--3">
-            <div class="card">
-              <div class="card__header">
-                <div class="avatar">
-                  <img
-                    class="avatar__photo"
-                    src={quote.avatar} />
-                  <div class="avatar__intro">
-                    <div class="avatar__name">{quote.name}</div>
-                    <small class="avatar__subtitle">
-                      {quote.title}
-                    </small>
-                  </div>
-                </div>
-              </div>
-              <div class="card__body">
-                <p> {quote.quote} </p>
-              </div>
-            </div>
-          </div>
-          ))}
-        </div>
-      </div>
-    </section>
-  );
-}
 
 export default function Home() {
   const {siteConfig} = useDocusaurusContext();
