+++ src/pages/index.module.css
@@ -41,7 +41,3 @@
   width: 100%;
   height: 20em;
 }
-
-.quotes h1 {
-  text-align: center;
-}
